Software Development For Computer Networks And Telecommunications
Brief Summary :
Consists of 2 applications. An android application "TerminalApp" which runs on an Android Device (Android 6.0+) and a Java Application "edgeServer" which runs on a Desktop/Laptop and communicates with a mySQL Database. Simulation uses Android API(23 lvl), Java 11.0, SQL Server 5.7, JDBC connectorJ and MQTT 3.1 protocols.
This is a project simulating a network with 2 android devices(cars) and an edge server (laptop) where the android devices simulate movement inside a google map grid using SUMO to create geographical movement data, and a gaussian random distribution to generate RSSI and Throughput values. The edge server is trained with a set of data (in xml form from SUMO) and creates a heatmap based on RSSI and Throughput for the current map part. The Android devices communicate with edge server with 1 Hz frequency using the MQTT protocol (publish-subscribe protocol) sending their current simulation data. The edge server predicts the future (1 sec horizon) location and RSSI and Throughput values , replies back to the terminals and then stores the predicted data into the database using JDBC protocol. The android terminals use the predicted and real information and draw their movements on two google maps API fragments, and upon finishing transmission, the edge server outputs some statistics and creates a chart to illustrate the error of the predicted and real location distance divergence.